Group menus and backend fixes

+++ src/java/ResourcePackage/GroupsResource.java
@@ -14,8 +14,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.UriInfo;
 import javax.ws.rs.Produces;
@@ -65,8 +63,11 @@ public class GroupsResource {
                 if(authResult && myID == starterID){ 
                     User groupstarter = userdao.getUser(starterID);
                     int groupID = groupdao.createGroup(groupname, groupstarter);
-                    return Response.status(201).entity(groupID).build();
-                    
+                    if(groupID != 0){
+                        return Response.status(201).entity(groupID).build();
+                    }else{
+                        return Response.status(400 ).entity("Groupname is invalid.").build();
+                    }
                 }else{
                    return Response.status(401).entity("Authorization failed.").build(); 
                 }
@@ -92,30 +93,123 @@ public class GroupsResource {
     
     @PUT
     @Path("/{groupid}")
-    public Response updateGroup(@PathParam("groupid") int groupid) {
-        return null;
-        //TODO
-    }
-    
-    @DELETE
-    @Path("/{groupid}")
-    public Response deleteGroup(@PathParam("groupid") int groupid) {
-        return null;
-        //TODO
-    }
-    //TODO Only admins can add users
-    @POST
-    @Path("/{groupid}/users")
-    public Response addUserToGroup(@PathParam("groupid") int groupID, @FormParam("userID") int userID, @Context HttpHeaders headers) {
-         /*if(groupdao.getGroups().containsKey(groupID) && userdao.getUsers().containsKey(userID)){
+    public Response updateGroup(@PathParam("groupid") int groupid,@FormParam("newname") String newname, @Context HttpHeaders headers) {
+        if(groupdao.getGroups().containsKey(groupid)){
             if(headers.getRequestHeaders().keySet().contains("authorization")){
                 String authCredentials = headers.getRequestHeader("authorization").get(0);
                 boolean authResult = userdao.authenticateUser(authCredentials);
                 int myID = userdao.getUserByName(userdao.decodeUsername(authCredentials)).getUserID();
-                if(authResult && myID == starterID){ 
-                    User groupstarter = userdao.getUser(starterID);
-                    int groupID = groupdao.createGroup(groupname, groupstarter);
-                    return Response.status(201).entity(groupID).build();
+                
+                if(authResult){
+                    if(groupdao.getGroup(groupid).getGroupAdmins().containsKey(myID)){
+                        String changeGroupname = groupdao.updateGroup(groupid, newname);
+                        if(!changeGroupname.equals("Failed")){
+                            return Response.status(200).entity("Groupname changed.").build(); 
+                        }else{
+                            return Response.status(400).entity("Groupname is invalid.").build(); 
+                        }
+                    }else{
+                       return Response.status(401).entity("You are not a group admin.").build();  
+                    }
+                }else{
+                   return Response.status(401).entity("Authorization failed.").build(); 
+                }
+            }else{
+                return Response.status(401).entity("Authorization failed.").build();
+            }
+        }else{
+            return Response.status(404).entity("Group not found.").build();
+        } 
+    }
+    
+    @DELETE
+    @Path("/{groupid}")
+    public Response deleteGroup(@PathParam("groupid") int groupid, @Context HttpHeaders headers) {
+      if(groupdao.getGroups().containsKey(groupid)){
+            if(headers.getRequestHeaders().keySet().contains("authorization")){
+                String authCredentials = headers.getRequestHeader("authorization").get(0);
+                boolean authResult = userdao.authenticateUser(authCredentials);
+                int myID = userdao.getUserByName(userdao.decodeUsername(authCredentials)).getUserID();
+                if(authResult){
+                    if(groupdao.getGroup(groupid).getGroupAdmins().containsKey(myID)){
+                        //Removing the group from all users within the group
+                        Group g = groupdao.getGroup(groupid);
+                        ArrayList<User> removedusers = new ArrayList<User>();
+                        for(Map.Entry<Integer,User> entry : g.getUserlist().entrySet()){
+                            removedusers.add(entry.getValue());
+                            //groupdao.removeUserFromGroup(groupid, entry.getValue());
+                        } 
+                        for(User u : removedusers){
+                            groupdao.removeUserFromGroup(groupid, u);
+                            if(u.getGrouplist().containsKey(groupid)){
+                                System.out.println("Vituiks meni "+u.getUserID()+" "+groupid);
+                            }
+                        }
+                        
+                        System.out.println(groupdao.getGroup(groupid).getUserlist().size());
+                        return Response.status(200).entity("Group removed").build(); 
+                    }else{
+                       return Response.status(401).entity("You are not a group admin.").build();  
+                    }
+                }else{
+                   return Response.status(401).entity("Authorization failed.").build(); 
+                }
+            }else{
+                return Response.status(401).entity("Authorization failed.").build();
+            }
+        }else{
+            return Response.status(404).entity("Group not found.").build();
+        }  
+    }
+    
+    //Leave group
+    @DELETE
+    @Path("/{groupid}/leave")
+    public Response leaveGroup(@PathParam("groupid") int groupid, @Context HttpHeaders headers) {
+       if(groupdao.getGroups().containsKey(groupid)){
+            if(headers.getRequestHeaders().keySet().contains("authorization")){
+                String authCredentials = headers.getRequestHeader("authorization").get(0);
+                boolean authResult = userdao.authenticateUser(authCredentials);
+                int myID = userdao.getUserByName(userdao.decodeUsername(authCredentials)).getUserID();
+                
+                if(authResult){
+                    if(groupdao.getGroup(groupid).getUserlist().containsKey(myID)){
+                        groupdao.removeUserFromGroup(groupid, userdao.getUser(myID));
+                        return Response.status(200).entity("Succesfully left the group.").build(); 
+                    }else{
+                       return Response.status(401).entity("You are not part of this group.").build();  
+                    }
+                }else{
+                   return Response.status(401).entity("Authorization failed.").build(); 
+                }
+            }else{
+                return Response.status(401).entity("Authorization failed.").build();
+            }
+        }else{
+            return Response.status(404).entity("Group not found.").build();
+        }  
+    }
+    
+    @POST
+    @Path("/{groupid}/users")
+    public Response addUserToGroup(@PathParam("groupid") int groupID, @FormParam("userID") int userID, @Context HttpHeaders headers) {
+         if(groupdao.getGroups().containsKey(groupID) && userdao.getUsers().containsKey(userID)){
+            if(headers.getRequestHeaders().keySet().contains("authorization")){
+                String authCredentials = headers.getRequestHeader("authorization").get(0);
+                boolean authResult = userdao.authenticateUser(authCredentials);
+                int myID = userdao.getUserByName(userdao.decodeUsername(authCredentials)).getUserID();
+                
+                if(authResult){
+                    if(groupdao.getGroup(groupID).getGroupAdmins().containsKey(myID)){
+                        if(!groupdao.getGroup(groupID).getUserlist().containsKey(userID)){
+                            groupdao.addUserToGroup(groupID, userdao.getUser(userID));
+                            return Response.status(200).entity("User "+userID+" added to group "+groupID).build();
+                        }else{
+                            return Response.status(409).entity("User "+userID+" is already in this group.").build(); 
+                        }
+                    }else{
+                       return Response.status(401).entity("You are not a group admin.").build();  
+                    }
                     
                 }else{
                    return Response.status(401).entity("Authorization failed.").build(); 
@@ -126,18 +220,99 @@ public class GroupsResource {
         }else{
             return Response.status(404).entity("User or group not found.").build();
         } 
-        */
-         
-        if(groupdao.getGroups().containsKey(groupID) && userdao.getUsers().containsKey(userID)){
-            if(!groupdao.getGroup(groupID).getUserlist().containsKey(userID)){
-                groupdao.addUserToGroup(groupID, userdao.getUser(userID));
-                return Response.status(200).entity("User "+userID+" added to group "+groupID).build();
+    }
+    @GET
+    @Path("/{groupid}/admins")
+    @Produces(MediaType.APPLICATION_XML)
+     public Response getGroupAdmins(@PathParam("groupid") int groupID,@Context HttpHeaders headers) {
+         if(groupdao.getGroups().containsKey(groupID)){
+            
+                if(headers.getRequestHeaders().keySet().contains("authorization")){
+                    String authCredentials = headers.getRequestHeader("authorization").get(0);
+                    boolean authResult = userdao.authenticateUser(authCredentials);
+                    //Getting current userID
+                    int myID = userdao.getUserByName(userdao.decodeUsername(authCredentials)).getUserID();
+                    
+                    if(authResult){
+                        if(groupdao.getGroup(groupID).getUserlist().containsKey(myID)){
+                            Group g = groupdao.getGroup(groupID);
+                            ArrayList returnarray = new ArrayList<User>();
+                            for(Map.Entry<Integer,User> entry : g.getGroupAdmins().entrySet()){
+                                returnarray.add(entry.getValue());
+                            } 
+                            GenericEntity<List<User>> entity = new GenericEntity<List<User>>(returnarray) {};
+                            return Response.status(200).entity(entity).build();
+                            
+                        }else{
+                            return Response.status(401).entity("You are not part of this group.").build();
+                        }
+                    }else{
+                       return Response.status(401).entity("Authorization failed.").build(); 
+                    }
+                }else{
+                    return Response.status(401).entity("Authorization failed.").build();
+                }
+        }else{
+            return Response.status(404).entity("Group not found.").build();
+        }
+    }
+    @POST
+    @Path("/{groupid}/admins")
+     public Response addGroupAdmin(@PathParam("groupid") int groupID, @FormParam("userID") int userID,@Context HttpHeaders headers) {
+         if(groupdao.getGroups().containsKey(groupID) && userdao.getUsers().containsKey(userID)){
+            if(headers.getRequestHeaders().keySet().contains("authorization")){
+                String authCredentials = headers.getRequestHeader("authorization").get(0);
+                boolean authResult = userdao.authenticateUser(authCredentials);
+                int myID = userdao.getUserByName(userdao.decodeUsername(authCredentials)).getUserID();
+                if(authResult){
+                    if(groupdao.getGroup(groupID).getGroupAdmins().containsKey(myID)){
+                        if(!groupdao.getGroup(groupID).getGroupAdmins().containsKey(userID)){
+                            groupdao.promoteToAdmin(groupID, userdao.getUser(userID));
+                            return Response.status(200).entity("User "+userID+" added to group "+groupID).build();
+                        }else{
+                            return Response.status(409).entity("User "+userID+" is already an admin").build(); 
+                        }
+                    }else{
+                       return Response.status(401).entity("You are not a group admin.").build();  
+                    }
+                }else{
+                   return Response.status(401).entity("Authorization failed.").build(); 
+                }
             }else{
-                return Response.status(409).entity("User "+userID+" is already in this group.").build(); 
+                return Response.status(401).entity("Authorization failed.").build();
             }
         }else{
-            return Response.status(400).entity("Invalid group or user ID").build(); 
-        }
+            return Response.status(404).entity("User or group not found.").build();
+        } 
+    }
+    @DELETE
+    @Path("/{groupid}/admins")
+    public Response removeGroupAdmin(@PathParam("groupid") int groupID, @PathParam("userID") int userID,@Context HttpHeaders headers) {
+         if(groupdao.getGroups().containsKey(groupID) && userdao.getUsers().containsKey(userID)){
+            if(headers.getRequestHeaders().keySet().contains("authorization")){
+                String authCredentials = headers.getRequestHeader("authorization").get(0);
+                boolean authResult = userdao.authenticateUser(authCredentials);
+                int myID = userdao.getUserByName(userdao.decodeUsername(authCredentials)).getUserID();
+                if(authResult){
+                    if(groupdao.getGroup(groupID).getGroupAdmins().containsKey(myID)){
+                        if(groupdao.getGroup(groupID).getGroupAdmins().containsKey(userID)){
+                            groupdao.demoteFromADmin(groupID, userdao.getUser(userID));
+                            return Response.status(200).entity("User "+userID+" added to group "+groupID).build();
+                        }else{
+                            return Response.status(409).entity("User "+userID+" is not an admin").build(); 
+                        }
+                    }else{
+                       return Response.status(401).entity("You are not a group admin.").build();  
+                    }
+                }else{
+                   return Response.status(401).entity("Authorization failed.").build(); 
+                }
+            }else{
+                return Response.status(401).entity("Authorization failed.").build();
+            }
+        }else{
+            return Response.status(404).entity("User or group not found.").build();
+        } 
     }
     
     @GET
@@ -180,13 +355,35 @@ public class GroupsResource {
     //TODO admin
     @DELETE
     @Path("/{groupid}/users/{userid}")
-    public Response removeUserFromGroup(@PathParam("groupid") int groupID, @PathParam("userid") int userID) {
-        if(groupdao.getGroups().containsKey(groupID) && userdao.getUsers().containsKey(userID)){
-            groupdao.removeUserFromGroup(groupID,userdao.getUser(userID));
-            return Response.status(200).entity("User removed from group.").build();
+    public Response removeUserFromGroup(@PathParam("groupid") int groupID, @PathParam("userid") int userID, @Context HttpHeaders headers) {
+        
+         if(groupdao.getGroups().containsKey(groupID) && userdao.getUsers().containsKey(userID)){
+            if(headers.getRequestHeaders().keySet().contains("authorization")){
+                String authCredentials = headers.getRequestHeader("authorization").get(0);
+                boolean authResult = userdao.authenticateUser(authCredentials);
+                int myID = userdao.getUserByName(userdao.decodeUsername(authCredentials)).getUserID();
+                
+                if(authResult){
+                    if(groupdao.getGroup(groupID).getGroupAdmins().containsKey(myID)){
+                        if(groupdao.getGroup(groupID).getUserlist().containsKey(userID)){
+                            groupdao.removeUserFromGroup(groupID,userdao.getUser(userID));
+                            return Response.status(200).entity("User removed from group.").build();
+                        }else{
+                            return Response.status(409).entity("User "+userID+" is not in this group.").build(); 
+                        }
+                    }else{
+                       return Response.status(401).entity("You are not a group admin.").build();  
+                    }
+                    
+                }else{
+                   return Response.status(401).entity("Authorization failed.").build(); 
+                }
+            }else{
+                return Response.status(401).entity("Authorization failed.").build();
+            }
         }else{
-            return Response.status(400).entity("Invalid group or user ID").build(); 
-        }
+            return Response.status(404).entity("User or group not found.").build();
+        } 
     }
     //Useless?
     @GET
